Move ReportCategory to its own class for reuse in AdminReports entity
